% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download.R
\name{DownloadL8S2}
\alias{DownloadL8S2}
\title{DownloadL8S2}
\usage{
DownloadL8S2(
  Sites,
  StartDate,
  EndDate,
  Indices = c("NDVI", "EVI"),
  Radius = 50,
  Outname = "Out.csv"
)
}
\arguments{
\item{Sites}{a dataframe with 3 columns : ID, x, y}

\item{StartDate}{Start date as a character}

\item{EndDate}{End date as a character}

\item{Indices}{vector of VI indices to be calculated}

\item{Radius}{Numeric Buffer radius}

\item{Outname}{charcter name with csv extension}
}
\value{

}
\description{
DownloadL8S2
}
\examples{
\dontrun{
mysites <- data.frame(x= c(-88.20,-88.20),
y= c(40.06,42.06),
ID= c('EnergyFarm',"RandomFarm")
)
RS <- DownloadL8S2(mysites,
                  '2020-01-01',
                  '2021-01-01')

}
}
